pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rune-agent"
version = "0.1.0"
description = "Rune — Local-First AI Coding Agent for Your Terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Deng Ming" },
]
keywords = ["ai", "agent", "coding", "local-llm", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
rune = "rune.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
